Did some cleaning and added comments
